fix(msvc): silence C4244 narrowings at the client and flen-test sites
zbc_client.c:509 -- zbc_read_native_int() returns intptr_t to cover the
host-side case where int_size > sizeof(int); on the client int_size
equals sizeof(int) by construction, so narrow with an explicit (int)
cast and document the invariant.

test_ansi_insecure.c:145, 643 -- file lengths come back as intmax_t;
hold them in intmax_t locals instead of an int. The second site adds a
separate flen_result variable so the strlen/read branch can keep its
int len.

+++ src/c/zbc_client.c
@@ -506,7 +506,10 @@ int zbc_parse_response(zbc_response_t *response, const uint8_t *buf,
       found_retn = 1;
       /* RETN: result[int_size] + errno[4] + optional DATA sub-chunk */
       if (size >= (size_t)int_size + ZBC_RETN_ERRNO_SIZE) {
-        response->result = zbc_read_native_int(chunk_data, int_size, endian);
+        /* int_size equals sizeof(int) on the client by construction; the
+         * cast cannot lose data here. Host-side parsing keeps the wider
+         * intptr_t (see zbc_protocol.h's parsed-message struct). */
+        response->result = (int)zbc_read_native_int(chunk_data, int_size, endian);
         response->error_code = (int)ZBC_READ_U32_LE(chunk_data + int_size);
 
         /* Check for DATA sub-chunk within RETN */
